installer.wxs: do not install XenNet by default
XenNet is now disabled in "Typical" installation, and unchecked in the list of
components in the "Custom" installation.

As described in #18, the use of XenNet breaks our auto-net config using our
DHCP server in a stub domain (because the traffic is now not routed through
stubdom anymore).

I have determined that the performance impact when using qemu-emulated NIC vs.
XenNet-provided NIC is negligible for a user using a WiFi network.
Specifically, the transfer rate (copying a large file to Win7 HVM)
over an Ethernet cable achieved a rate or around 20MB/s with XenNet, while 12
MB/s with qemu NIC. Taking into account that 12MB/s (100Mps) is the upper limit
on the fastest 802.11n, it doesn't seem like a much compromise... And users who
really feel they need it, can always choose "Custom" installation type and
install this driver (and configure the networking manually for the guest).

On a side note: the limiting factor in case I was using the cable ethernet seeme
d to be the tasks in firewallvm and netvm -- I could see the CPU load in those VMs to be around 25%, which corresponds to one out of my 4 course being fully uti
lized, and forming a bottleneck.