xenpci: present removable devices in Unplug or Eject Hardware program (#22 pro)
Devices with EjectSupported = True and SurpriseRemoveOK = False are subject to
Unplug or Eject Hardware program. Set it based on 'removable' xenstore
flag.

This can be done only by XenPci (bus driver), not driver itself, but currently
only block devices have 'removable' flag. This code ignore this fact - it
implements 'removable' flag for all devices, perhaps it will be useful in the
future.